Move resistance example scripts to archive/ with descriptive names
Renames:
- run_resistance_demo.py -> run_resistance_program_demo.py
- run_resistance_scenario_driven.py -> run_resistance_spec_validation.py
- run_resistance_strainaware_int.py -> run_resistance_tpt_tradeoff.py
- run_resistance_critical_paths.py (unchanged, moved to archive/)
- run_resistance.py, run_resistance2.py (moved to archive/)
