+++ tbsim_examples/archive/resistance/run_resistance_spec_validation.py
