+++ tbsim_examples/archive/resistance/run_resistance_tpt_tradeoff.py
